Design 1: Toolbar above Brick and Tutorial Editor but still two seperate blocks
