+++ frontend/src/app/views/brick_settings/content_group.rs
@@ -8,10 +8,6 @@ use yew::prelude::*;
 #[cfg(target_arch = "wasm32")]
 const ICON_ADD: &str = include_str!("../../../res/add.svg");
 #[cfg(target_arch = "wasm32")]
-const ICON_UPLOAD: &str = include_str!("../../../res/upload.svg");
-#[cfg(target_arch = "wasm32")]
-const ICON_FILE_JSON: &str = include_str!("../../../res/file_json.svg");
-#[cfg(target_arch = "wasm32")]
 const ICON_IMAGE: &str = include_str!("../../../res/image.svg");
 #[cfg(target_arch = "wasm32")]
 const ICON_FILE_PNG: &str = include_str!("../../../res/file_png.svg");
@@ -22,8 +18,6 @@ pub struct ContentGroupProps {
     pub content: String,
     pub on_content_input: Callback<InputEvent>,
     pub on_add_to_tutorial: Callback<MouseEvent>,
-    pub on_import_json: Callback<MouseEvent>,
-    pub on_export_json: Callback<MouseEvent>,
     pub on_save_svg: Callback<MouseEvent>,
     pub on_save_png: Callback<MouseEvent>,
 }
@@ -44,16 +38,6 @@ pub fn content_group(props: &ContentGroupProps) -> Html {
                     icon={Html::from_html_unchecked(AttrValue::from(ICON_ADD))}
                     title="Add to tutorial"
                     onclick={props.on_add_to_tutorial.clone()}
-                />
-                <IconButton
-                    icon={Html::from_html_unchecked(AttrValue::from(ICON_UPLOAD))}
-                    title="Import JSON"
-                    onclick={props.on_import_json.clone()}
-                />
-                <IconButton
-                    icon={Html::from_html_unchecked(AttrValue::from(ICON_FILE_JSON))}
-                    title="Export JSON"
-                    onclick={props.on_export_json.clone()}
                 />
                 <IconButton
                     icon={Html::from_html_unchecked(AttrValue::from(ICON_IMAGE))}
