+++ frontend/src/app/views/brick_settings/mod.rs
@@ -62,25 +62,7 @@ pub fn brick_settings_view(props: &BrickSettingsViewProps) -> Html {
     let color_name = current.color_scheme.name.clone();
     let selected_color = current.color_scheme.clone();
 
-    let on_import_json = {
-        let dispatcher = props.dispatcher.clone();
-        Callback::from(move |_| {
-            let dispatcher = dispatcher.clone();
-            utility::upload_json(Callback::from(move |text: String| {
-                dispatcher.dispatch(StateAction::LoadJson(text));
-            }));
-        })
-    };
 
-    let on_export_json = {
-        let brick = props.brick.clone();
-        Callback::from(move |_| {
-            let json = brick.to_string();
-            if let Err(e) = utility::download_json(&json, "brick.json") {
-                web_sys::console::error_1(&format!("Export error: {e}").into());
-            }
-        })
-    };
 
     let on_save_svg = {
         let brick = props.brick.clone();
@@ -110,8 +92,6 @@ pub fn brick_settings_view(props: &BrickSettingsViewProps) -> Html {
                 content={content_val}
                 on_content_input={on_content_input}
                 on_add_to_tutorial={props.on_add_to_tutorial.clone()}
-                on_import_json={on_import_json}
-                on_export_json={on_export_json}
                 on_save_svg={on_save_svg}
                 on_save_png={on_save_png}
             />
