+++ frontend/styles/page.css
@@ -1,6 +1,43 @@
 .page {
     display: flex;
+    flex-direction: column;
     height: 100vh;
+    overflow: hidden;
+}
+
+.page__toolbar {
+    display: flex;
+    align-items: center;
+    gap: var(--gap);
+    padding: 8px var(--gap);
+    border-bottom: 1px solid var(--color-border);
+    background: var(--color-surface);
+    flex: 0 0 auto;
+    flex-wrap: wrap;
+}
+
+.page__toolbar-group {
+    display: flex;
+    align-items: center;
+    gap: 6px;
+    padding: 4px 8px;
+    border: 1px solid var(--color-border);
+    border-radius: var(--radius);
+    background: var(--color-surface-raised);
+}
+
+.page__toolbar-label {
+    font-size: 11px;
+    font-weight: 600;
+    text-transform: uppercase;
+    letter-spacing: 0.06em;
+    color: var(--color-text-muted);
+}
+
+.page__content {
+    display: flex;
+    flex: 1 1 0;
+    min-height: 0;
     overflow: hidden;
 }
